VarArrays can now be resized dynamically.
If the array is constructed with an initial size given, it will be
allocated in the space. If the array is resized, the new array will be
allocated on the heap instead. When copying an array, the copy is
always allocated in the space (but can again be resized).

This should ensure that static allocation is just as efficient as
before. As resize is typically only used for setting up the model, the
array will only be allocated on the heap for setup, not for any copies.

The only overhead is that VarArray now needs a destructor that checks
whether allocation happened on the heap, and if so frees the memory.



git-svn-id: https://svn.gecode.org/svn/gecode/trunk@4595 64335634-5103-0410-b293-fc3d331e086d